pyproject.toml
[tool.poetry]
name = "djangoSite"
version = "0.1.0"
description = ""
authors = ["ZwickVitaly <zwickvitaly@gmail.com>"]


[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
djangorestframework = "^3.15.1"
django-filter = "^24.2"
django-debug-toolbar = "^4.3.0"
pillow = "^10.3.0"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
